Añadí la funcionalidad para cambiar la función de agregación de las pivot tables

+++ python-app/lib/pivot_table/recalculate.py
@@ -5,6 +5,8 @@ from lib.pivot_table.ordered_filters import find_filter
 from lib.pivot_table.render_bare_preview_of_pivot_table import render_bare_preview_of_pivot_table
 from lib.data_filter.is_valid_filter import is_valid_filter
 from lib.data_frame.data_frame_io import import_data_frame
+from lib.slide.render_preview import render_preview
+from lib.slide.delete_preview import delete_preview
 
 from models.pivot_table.self import PivotTable
 from models.pivot_table.data_filter.self import DataFilter
@@ -37,13 +39,11 @@ def fix_charts(filters: list[DataFilter], filters_order: list[PivotTable]) -> No
         new_chart.charting_mode = ChartingMode.CHART
     
 
-def recalculate(report: Report, pivot_table: PivotTable, preloaded_data_frame: pandas.DataFrame | None = None):
+def recalculate(root_directory: str, pivot_table: PivotTable, preloaded_data_frame: pandas.DataFrame | None = None):
     '''
     Re calculates the data of the pivot table based on its filters. It modifies the pivot table in place.
     Also, it creates a new preview
     '''
-    root_directory = report.root_directory
-
     data_frame = preloaded_data_frame
     if data_frame is None:
         data_frame = import_data_frame(file_path=pivot_table.source.merged_file, key=pivot_table.identifier)
@@ -60,8 +60,12 @@ def recalculate(report: Report, pivot_table: PivotTable, preloaded_data_frame: p
         aggregate_function=pivot_table.aggregate_function
         )
     
-    if os.path.exists(pivot_table.bare_preview):
+    pivot_table.data = new_data
+    
+    if pivot_table.bare_preview is not None and os.path.exists(pivot_table.bare_preview):
         os.remove(pivot_table.bare_preview)
 
-    pivot_table.data = new_data
     render_bare_preview_of_pivot_table(root_directory=root_directory, pivot_table=pivot_table)
+
+    delete_preview(slide=pivot_table)
+    render_preview(root_directory=root_directory, slides=pivot_table)
